create folder for Properties-Stuff and move Manager in it
